+++ tests/testthat/test-wb_functions.R
@@ -34,7 +34,7 @@ test_that("wb_to_df", {
   expect_equal(convert_to_excel_date(df = exp["Var5"], date1904 = FALSE),
                got["Var5"])
 
-  expect_warning(convertToExcelDate(df = exp["Var5"], date1904 = FALSE), "deprecated")
+  expect_error(expect_warning(convertToExcelDate(df = exp["Var5"], date1904 = FALSE), "deprecated"),"defunct")
 
   # return the underlying Excel formula instead of their values
   got <- wb_to_df(wb1, showFormula = TRUE)